requirements.txt
awsebcli==3.15.3
blessed==1.16.1
boto3==1.10.6
botocore==1.12.253
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.6
django-cors-headers==3.2.0
django-extensions==2.2.5
django-rest-framework==0.1.0
django-storages==1.7.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.0
future==0.16.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==2.6.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==19.2
pathspec==0.5.9
Pillow==6.2.1
psycopg2-binary==2.8.4
pygraphviz==1.5
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
rest-social-auth==2.2.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
